backend/requirements.txt
fastapi==0.115.0
pydantic[standard]==2.9.2
python-multipart==0.0.12
mangum==0.19.0
numpy==1.26.0
opencv-python-headless==4.10.0.84
Pillow==10.0.0
requests==2.27.0
httpcore==1.0.0
uvicorn==0.40.0